Moved model function in json_formatter to model
