+++ json_formatter.py
@@ -18,42 +18,6 @@ class JsonFormatter(object):
 
         return search_word_json
 
-    def load_search_result(self, path = 'json_data/search_1135_ja.json'):
-        f = open(path, 'r')
-        search_result_read = json.load(f)
-        return search_result_read
-
-
-    #  twitter search shouldn't be triggerd here. But needs for tweet_json_init
-    def json_init(self, search_word_dict):
-
-        search_result = {}
-        tweet_list_temp = {}
-        tweet_list_json = {}
-
-        twitter_search = TwitterSearch()
-
-        for i in range(len(search_word_dict)):
-
-            if search_word_dict[i] == "":
-                continue
-
-            else:
-                search_result[str(i)] = twitter_search.search(search_word_dict[i])
-
-            # tweet_list_json[str(i)] = "i_init"
-
-            if search_word_dict[i] == "":
-                continue
-
-            else:
-                tweet_list_json[str(i)] = {}
-
-                for j in range(len(search_result[str(i)]["statuses"])):
-                    tweet_list_json[str(i)][str(j)] = "j_init"
-
-        return search_result, tweet_list_temp, tweet_list_json
-
 
 
     def tw_json_to_html_json(self, search_word_dict, search_result, tweet_list_temp,tweet_list_json):
@@ -111,13 +75,6 @@ class JsonFormatter(object):
 
 
 
-    def save_result_tweet(self, path = 'json_data/result_tweet_json.json', tweet_list_json = {}):
-        f = open(path, 'w')
-        json.dump(tweet_list_json, f, indent=4, ensure_ascii=False)
-        print("result_tweet_json saved")
-
-
-
     def print_json(self, json, search_word_dict):
         for i in range(len(json)):
             if search_word_dict[i] == "":
@@ -131,6 +88,3 @@ class JsonFormatter(object):
                     print(json[str(i)][str(j)]["u_name"])
                     print(json[str(i)][str(j)]["text"])
                     print("")
-
-
-
